requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
py-mon==2.0.5
requests==2.32.3
urllib3==2.2.2
watchdog==5.0.0
Werkzeug==3.0.4

PyYAML~=6.0.2